used moshi to convert json to PictureOfDay object
just notice the PictureOfDay data class, anotated with json.
Had to delete my manual process of converting the image of the day
json to my custom class.

did not know how to add two converters (Scalar and Moshi) to the same
retrofit service so that i could distinguish between the picture of the
day json and the near earth asteroid json, so just made two retrofit
services for each use case.